refactor: topbar 컴포넌트 모달 분리 및 최적화 작업
